// src/app/models/verb-sets/id-drill-options.model.ts
export interface IdDrillOptions {
    engAudio: string;
    tenseAudio: string;
    pronounAudio: string;
    arabicAudio: string;
    playCountA: number;
    betweenA: number;
    beforeB: number;
    betweenB: number;
    beforeRepeat: number;
    playCountRepeatA: number;
    spacerSound: string;
}